Add short to integer converter to graphitron-example-db

See merge request graphitron/graphitron!67

+++ graphitron-example/graphitron-example-service/src/main/java/no/sikt/graphitron/example/service/conditions/StaffConditions.java
@@ -13,6 +13,6 @@ public class StaffConditions {
     }
 
     public static Condition withAddressIdInCity300(Address address, Integer addressId) {
-        return addressId == null ? noCondition() : address.CITY_ID.eq((short) 300);
+        return addressId == null ? noCondition() : address.CITY_ID.eq(300);
     }
 }
